Node host: extract withTimeout helper
