+++ src/node-host/runner.ts
@@ -45,6 +45,7 @@ import { detectMime } from "../media/mime.js";
 import { GATEWAY_CLIENT_MODES, GATEWAY_CLIENT_NAMES } from "../utils/message-channel.js";
 import { VERSION } from "../version.js";
 import { ensureNodeHostConfig, saveNodeHostConfig, type NodeHostGatewayConfig } from "./config.js";
+import { withTimeout } from "./with-timeout.js";
 
 type NodeHostRunOptions = {
   gatewayHost: string;
@@ -273,40 +274,6 @@ async function ensureBrowserControlService(): Promise<void> {
     }
   })();
   return browserControlReady;
-}
-
-async function withTimeout<T>(
-  work: (signal: AbortSignal | undefined) => Promise<T>,
-  timeoutMs?: number,
-  label?: string,
-): Promise<T> {
-  const resolved =
-    typeof timeoutMs === "number" && Number.isFinite(timeoutMs)
-      ? Math.max(1, Math.floor(timeoutMs))
-      : undefined;
-  if (!resolved) {
-    return await work(undefined);
-  }
-  const abortCtrl = new AbortController();
-  const timeoutError = new Error(`${label ?? "request"} timed out`);
-  const timer = setTimeout(() => abortCtrl.abort(timeoutError), resolved);
-
-  let abortListener: (() => void) | undefined;
-  const abortPromise: Promise<never> = abortCtrl.signal.aborted
-    ? Promise.reject(abortCtrl.signal.reason ?? timeoutError)
-    : new Promise((_, reject) => {
-        abortListener = () => reject(abortCtrl.signal.reason ?? timeoutError);
-        abortCtrl.signal.addEventListener("abort", abortListener, { once: true });
-      });
-
-  try {
-    return await Promise.race([work(abortCtrl.signal), abortPromise]);
-  } finally {
-    clearTimeout(timer);
-    if (abortListener) {
-      abortCtrl.signal.removeEventListener("abort", abortListener);
-    }
-  }
 }
 
 function isProfileAllowed(params: { allowProfiles: string[]; profile?: string | null }) {
